Removed the concept of a time slice. The data is now always split in integration slices.
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@615 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/buffer.h
@@ -136,7 +136,7 @@ public:
         if( boost::is_pod<T>::value ) {
           memcpy(buffer_, oldbuffer, sizeof(T)*min_size);
         } else {
-          for (int i=0; i<min_size; i++) {
+          for (size_t i=0; i<min_size; i++) {
             buffer_[i] = oldbuffer[i];
           }
         }
